requirements.txt
scikit-learn==1.3.0
snowflake-connector-python==3.0.4
snowflake-snowpark-python==1.5.1
snowflake-snowpark-python[pandas]==1.5.1
python-dotenv==1.0.0
pandas==1.4.1
PyYAML==6.0